working on deleting card image

+++ DataAccess/Migrations/AppDbContextModelSnapshot.cs
@@ -80,11 +80,11 @@ namespace DataAccess.Migrations
 
             modelBuilder.Entity("Entities.CardImage", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"), 1L, 1);
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"), 1L, 1);
 
                     b.Property<long?>("CardId")
                         .HasColumnType("bigint");
